build fix
svn:sync-xref-src-repo-rev: r112258

+++ src/VBox/Runtime/tools/RTSignTool.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTSignTool.cpp 64850 2016-12-13 14:58:34Z knut.osmundsen@oracle.com $ */
+/* $Id: RTSignTool.cpp 64858 2016-12-13 17:54:43Z noreply@oracle.com $ */
 /** @file
  * IPRT - Signing Tool.
  */
@@ -662,7 +662,7 @@ static int HandleShowExeWorkerPkcs7Decode(PSHOWEXEPKCS7 pThis)
 {
     RTERRINFOSTATIC     ErrInfo;
     RTASN1CURSORPRIMARY PrimaryCursor;
-    RTAsn1CursorInitPrimary(&PrimaryCursor, pThis->pbBuf, pThis->cbBuf, RTErrInfoInitStatic(&ErrInfo),
+    RTAsn1CursorInitPrimary(&PrimaryCursor, pThis->pbBuf, (uint32_t)pThis->cbBuf, RTErrInfoInitStatic(&ErrInfo),
                             &g_RTAsn1DefaultAllocator, 0, "WinCert");
 
     int rc = RTCrPkcs7ContentInfo_DecodeAsn1(&PrimaryCursor.Cursor, 0, &pThis->ContentInfo, "CI");
